efivars: move efi_set_variable() and friends to src/shared/
The EFI variable setters have no users inside src/basic/ or any other
code below src/shared/; every caller lives in a binary that links
libsystemd-shared. Move them into src/shared/efivars.c so that a
follow-up commit can teach them to persist writes on firmware with a
file-backed variable store, which requires ESP discovery via
find-esp.c — infrastructure that is not available from src/basic/.

The file keeps the efivars name across tiers, following the same
pattern as tmpfile-util and mkdir (see 74d392ed1bab ("tree-wide:
standardize header names across src/fundamental, src/basic and
src/shared")): the shared header re-exports the basic one and adds the
shared-only functions, targets pick up whichever tier their include
path resolves first, and no callers need to change.

No functional change.
