+++ src/basic/efivars.c
@@ -8,12 +8,9 @@
 #include <unistd.h>
 
 #include "alloc-util.h"
-#include "chattr-util.h"
 #include "efivars.h"
 #include "fd-util.h"
-#include "io-util.h"
 #include "log.h"
-#include "memory-util.h"
 #include "stat-util.h"
 #include "string-util.h"
 #include "time-util.h"
@@ -202,121 +199,6 @@ int efi_get_variable_path(const char *variable, char **ret) {
                 efi_tilt_backslashes(*ret);
 
         return r;
-}
-
-static int efi_verify_variable(const char *variable, uint32_t attr, const void *value, size_t size) {
-        _cleanup_free_ void *buf = NULL;
-        size_t n;
-        uint32_t a;
-        int r;
-
-        assert(variable);
-        assert(value || size == 0);
-
-        r = efi_get_variable(variable, &a, &buf, &n);
-        if (r < 0)
-                return r;
-
-        return a == attr && memcmp_nn(buf, n, value, size) == 0;
-}
-
-int efi_set_variable(const char *variable, const void *value, size_t size) {
-        static const uint32_t attr = EFI_VARIABLE_NON_VOLATILE|EFI_VARIABLE_BOOTSERVICE_ACCESS|EFI_VARIABLE_RUNTIME_ACCESS;
-
-        _cleanup_free_ struct var {
-                uint32_t attr;
-                char buf[];
-        } _packed_ *buf = NULL;
-        _cleanup_close_ int fd = -EBADF;
-        bool saved_flags_valid = false;
-        unsigned saved_flags;
-        int r;
-
-        assert(variable);
-        assert(value || size == 0);
-
-        /* size 0 means removal, empty variable would not be enough for that */
-        if (size > 0 && efi_verify_variable(variable, attr, value, size) > 0) {
-                log_debug("Variable '%s' is already in wanted state, skipping write.", variable);
-                return 0;
-        }
-
-        const char *p = strjoina("/sys/firmware/efi/efivars/", variable);
-
-        /* Newer efivarfs protects variables that are not in an allow list with FS_IMMUTABLE_FL by default,
-         * to protect them for accidental removal and modification. We are not changing these variables
-         * accidentally however, hence let's unset the bit first. */
-
-        r = chattr_full(AT_FDCWD, p,
-                        /* value= */ 0,
-                        /* mask= */ FS_IMMUTABLE_FL,
-                        /* ret_previous= */ &saved_flags,
-                        /* ret_final= */ NULL,
-                        /* flags= */ 0);
-        if (r < 0 && r != -ENOENT)
-                log_debug_errno(r, "Failed to drop FS_IMMUTABLE_FL flag from '%s', ignoring: %m", p);
-
-        saved_flags_valid = r >= 0;
-
-        if (size == 0) {
-                if (unlink(p) < 0) {
-                        r = -errno;
-                        goto finish;
-                }
-
-                return 0;
-        }
-
-        fd = open(p, O_WRONLY|O_CREAT|O_NOCTTY|O_CLOEXEC, 0644);
-        if (fd < 0) {
-                r = -errno;
-                goto finish;
-        }
-
-        buf = malloc(sizeof(uint32_t) + size);
-        if (!buf) {
-                r = -ENOMEM;
-                goto finish;
-        }
-
-        buf->attr = attr;
-        memcpy(buf->buf, value, size);
-
-        r = loop_write(fd, buf, sizeof(uint32_t) + size);
-        if (r < 0)
-                goto finish;
-
-        /* For some reason efivarfs doesn't update mtime automatically. Let's do it manually then. This is
-         * useful for processes that cache EFI variables to detect when changes occurred. */
-        if (futimens(fd, /* times= */ NULL) < 0)
-                log_debug_errno(errno, "Failed to update mtime/atime on %s, ignoring: %m", p);
-
-        r = 0;
-
-finish:
-        if (saved_flags_valid) {
-                int q;
-
-                /* Restore the original flags field, just in case */
-                if (fd < 0)
-                        q = chattr_path(p, saved_flags, FS_IMMUTABLE_FL);
-                else
-                        q = chattr_fd(fd, saved_flags, FS_IMMUTABLE_FL);
-                if (q < 0)
-                        log_debug_errno(q, "Failed to restore FS_IMMUTABLE_FL on '%s', ignoring: %m", p);
-        }
-
-        return r;
-}
-
-int efi_set_variable_string(const char *variable, const char *value) {
-        _cleanup_free_ char16_t *u16 = NULL;
-
-        u16 = utf8_to_utf16(value, SIZE_MAX);
-        if (!u16)
-                return -ENOMEM;
-
-        return efi_set_variable(variable, u16, (char16_strlen(u16) + 1) * sizeof(char16_t));
 }
 
 static int cache_efi_boot = -1;
