+++ src/basic/efivars.h
@@ -42,8 +42,6 @@
 int efi_get_variable(const char *variable, uint32_t *attribute, void **ret_value, size_t *ret_size);
 int efi_get_variable_string(const char *variable, char **ret);
 int efi_get_variable_path(const char *variable, char **ret);
-int efi_set_variable(const char *variable, const void *value, size_t size) _nonnull_if_nonzero_(2, 3);
-int efi_set_variable_string(const char *variable, const char *value);
 
 bool set_efi_boot(bool b);
 bool is_efi_boot(void);
@@ -61,14 +59,6 @@ static inline int efi_get_variable_string(const char *variable, char **ret) {
 }
 
 static inline int efi_get_variable_path(const char *variable, char **ret) {
-        return -EOPNOTSUPP;
-}
-
-static inline int efi_set_variable(const char *variable, const void *value, size_t size) {
-        return -EOPNOTSUPP;
-}
-
-static inline int efi_set_variable_string(const char *variable, const char *p) {
         return -EOPNOTSUPP;
 }
